Crate zng_ext_image

Image loading, rendering and cache.

Services

Services this extension provides.

IMAGES

Crate

This crate is part of the zng project.

Cargo Features

This crate provides 1 feature flag, not enabled by default.

"http"

Enable web image source.

Structs

ColorType

Basic info about a color model.

IMAGES

Image loading, cache and render service.

IMAGES_WINDOW

Images render window hook.

IMAGE_RENDER

Controls properties of the render window used by IMAGES.render.

ImageEntriesMode

Defines what images are decoded from multi image containers.

ImageEntry

State of an ImageVar.

ImageFormat

Represents an image codec capability.

ImageFormatCapability

Capabilities of an ImageFormat implementation.

ImageHash

A 256-bit hash for image entries.

ImageHasher

Hasher that computes a ImageHash.

ImageLimits

Limits for image loading and decoding.

ImageOptions

Options for IMAGES.image.

ImageRenderArgs

Arguments for the ImageSource::Render closure.

Enums

ImageCacheMode

Cache mode of IMAGES.

ImageDataFormat

Format of the image bytes.

ImageDownscaleMode

Defines how an image is downscaled after decoding.

ImageEntryKind

Kind of image container entry an image was decoded from.

ImageMaskMode

Defines how the A8 image mask pixels are to be derived from a source mask image.

ImageSource

The different sources of an image resource.

ImageSourceFilter

Represents a PathFilter and UriFilter.

PartialImageKind

Represents what kind of partial data was decoded.

Traits

ImageRenderWindowRoot

Implemented for the root window type.

ImageRenderWindowsService

Reference to a windows manager service that IMAGES can use to render images.

ImagesExtension

A custom extension for the IMAGES service.

Functions

render_retain

P If the render task is kept alive after a frame is produced, this is false by default meaning the image only renders once, if set to true the image will automatically update when the render widget requests a new frame.

Type Aliases

ImageVar

Represents an ImageEntry tracked by the IMAGES cache.

PathFilter

Represents an ImageSource::Read path request filter.

UriFilter

Represents an ImageSource::Download path request filter.