zng_ext_image
Enum ImageSource

pub enum ImageSource {
    Read(PathBuf),
    Download(Uri, Option<Txt>),
    Static(ImageHash, &'static [u8], ImageDataFormat),
    Data(ImageHash, Arc<Vec<u8>>, ImageDataFormat),
    Render(Arc<Box<dyn Fn(&ImageRenderArgs) -> Box<dyn ImageRenderWindowRoot> + Send + Sync>>, Option<ImageRenderArgs>),
    Image(ImageVar),
}

The different sources of an image resource.

Variants

Read(PathBuf)

A path to an image file in the file system.

Image equality is defined by the path, a copy of the image in another path is a different image.

Download(Uri, Option<Txt>)

A uri to an image resource downloaded using HTTP GET with an optional HTTP ACCEPT string.

If the ACCEPT line is not given, all image formats supported by the view-process backend are accepted.

Image equality is defined by the URI and ACCEPT string.

Static(ImageHash, &'static [u8], ImageDataFormat)

Static bytes for an encoded or decoded image.

Image equality is defined by the hash, it is usually the hash of the bytes but it does not need to be.

Data(ImageHash, Arc<Vec<u8>>, ImageDataFormat)

Shared reference to bytes for an encoded or decoded image.

Image equality is defined by the hash, it is usually the hash of the bytes but it does not need to be.

Inside IMAGES the reference to the bytes is held only until the image finishes decoding.

Render(Arc<Box<dyn Fn(&ImageRenderArgs) -> Box<dyn ImageRenderWindowRoot> + Send + Sync>>, Option<ImageRenderArgs>)

A boxed closure that instantiates a WindowRoot that draws the image.

Use the render or render_node functions to construct this variant.

The closure is set by the image widget user, the args is set by the image widget.

Image(ImageVar)

Already resolved (loaded or loading) image.

The image is passed-through, not cached.

Implementations

impl ImageSource

pub fn flood( size: impl Into<PxSize>, color: impl Into<Rgba>, ppi: Option<ImagePpi>, ) -> Self

New image data from solid color.

pub fn from_data(data: Arc<Vec<u8>>, format: ImageDataFormat) -> Self

New source from data.

pub fn from_static(data: &'static [u8], format: ImageDataFormat) -> Self

New source from static data.

pub fn hash128( &self, downscale: Option<ImageDownscale>, mask: Option<ImageMaskMode>, ) -> Option<ImageHash>

Returns the image hash, unless the source is Img.

pub fn hash128_data( data_hash: ImageHash, downscale: Option<ImageDownscale>, mask: Option<ImageMaskMode>, ) -> ImageHash

Compute hash for a borrowed Static or Data image.

pub fn hash128_read( path: &Path, downscale: Option<ImageDownscale>, mask: Option<ImageMaskMode>, ) -> ImageHash

Compute hash for a borrowed Read path.

pub fn hash128_download( uri: &Uri, accept: &Option<Txt>, downscale: Option<ImageDownscale>, mask: Option<ImageMaskMode>, ) -> ImageHash

Compute hash for a borrowed Download URI and HTTP-ACCEPT.

pub fn hash128_render( rfn: &Arc<Box<dyn Fn(&ImageRenderArgs) -> Box<dyn ImageRenderWindowRoot> + Send + Sync>>, args: &Option<ImageRenderArgs>, downscale: Option<ImageDownscale>, mask: Option<ImageMaskMode>, ) -> ImageHash

Compute hash for a borrowed Render source.

Pointer equality is used to identify the node closure.

impl ImageSource

pub fn render<F, R>(new_img: F) -> Self
where F: Fn(&ImageRenderArgs) -> R + Send + Sync + 'static, R: ImageRenderWindowRoot,

New image from a function that generates a headless window.

The function is called every time the image source is resolved and it is not found in the cache.

Examples

    |args| Window! {
        size = (500, 400);
        parent = args.parent;
        background_color = colors::GREEN;
        child = Text!("Rendered!");
    }
)

pub fn render_node<U, N>(render_mode: RenderMode, render: N) -> Self
where U: UiNode, N: Fn(&ImageRenderArgs) -> U + Send + Sync + 'static,

New image from a function that generates a new UiNode.

The function is called every time the image source is resolved and it is not found in the cache.

Note that the generated UiNode is not a child of the widget that renders the image, it is the root widget of a headless surface, not a part of the context where it is rendered. See IMAGES.render for more information.

Examples

ImageSource::render_node(
    RenderMode::Software,
    |_args| Container! {
        size = (500, 400);
        background_color = colors::GREEN;
        child = Text!("Rendered!");
    }
)