zng_ext_image

Enum ImageSourceFilter

pub enum ImageSourceFilter<U> {
    BlockAll,
    AllowAll,
    Custom(Arc<dyn Fn(&U) -> bool + Send + Sync>),
}

Represents a PathFilter and UriFilter.

Variants

BlockAll

Block all requests of this type.

AllowAll

Allow all requests of this type.

Custom(Arc<dyn Fn(&U) -> bool + Send + Sync>)

Custom filter, returns true to allow a request, false to block.

Implementations

impl<U> ImageSourceFilter<U>

pub fn custom(allow: impl Fn(&U) -> bool + Send + Sync + 'static) -> Self

New Custom filter.

pub fn and(self, other: Self) -> Self
where U: 'static,

Combine self with other, if they both are Custom, otherwise is BlockAll if any is BlockAll, else is AllowAll if any is AllowAll.

If both are Custom both filters must allow a request to pass the new filter.

pub fn or(self, other: Self) -> Self
where U: 'static,

Combine self with other, if they both are Custom, otherwise is AllowAll if any is AllowAll, else is BlockAll if any is BlockAll.

If both are Custom at least one of the filters must allow a request to pass the new filter.

pub fn allows(&self, item: &U) -> bool

Returns true if the filter allows the request.

impl ImageSourceFilter<PathBuf>

pub fn allow_dir(dir: impl AsRef<Path>) -> Self

Allow any file inside dir or sub-directories of dir.

pub fn allow_ext(ext: impl Into<OsString>) -> Self

Allow any path with the ext extension.

pub fn allow_current_dir() -> Self

Allow any file inside the env::current_dir or sub-directories.

Note that the current directory can be changed and the filter always uses the fresh current directory, use allow_dir to create a filter the always points to the current directory at the filter creation time.

pub fn allow_exe_dir() -> Self

Allow any file inside the current executable directory or sub-directories.

pub fn allow_res() -> Self

Allow any file inside the zng::env::res directory or sub-directories.

impl ImageSourceFilter<Uri>

pub fn allow_host(host: impl Into<Txt>) -> Self

Allow any file from the host site.