zng_ext_image

Struct ImageEntry

pub struct ImageEntry { /* private fields */ }

State of an ImageVar.

Implementations

impl ImageEntry

pub fn new_loading() -> Self

Create a dummy image in the loading state.

This is the same as calling new_error with an empty error.

pub fn new_error(error: Txt) -> Self

Create a dummy image in the error state.

If the error is empty the image is loading, not an error.

pub fn is_loading(&self) -> bool

Returns true if the is still acquiring or decoding the image bytes.

pub fn is_loaded(&self) -> bool

If the image has finished loading ok or due to error.

The image variable may still update after

pub fn is_error(&self) -> bool

If the image failed to load.

pub fn error(&self) -> Option<Txt>

Returns an error message if the image failed to load.

pub fn size(&self) -> PxSize

Pixel size of the image after it finishes loading.

Note that this value is set as soon as the header finishes decoding, but the pixels will only be set after it the entire image decodes.

If the view-process implements progressive decoding you can use partial_size and partial_pixels to use the partially decoded image top rows as it decodes.

pub fn partial_size(&self) -> Option<PxSize>

Size of partial_pixels.

Can be different from size if the image is progressively decoding.

pub fn partial_kind(&self) -> Option<PartialImageKind>

Kind of partial_pixels.

pub fn density(&self) -> Option<PxDensity2d>

Returns the image pixel density metadata if the image is loaded and the metadata was retrieved.

pub fn original_color_type(&self) -> ColorType

Image color type before it was converted to BGRA8 or A8.

pub fn color_type(&self) -> ColorType

Gets A8 for masks and BGRA8 for others.

pub fn is_opaque(&self) -> bool

Returns true if the image is fully opaque or it is not loaded.

pub fn is_mask(&self) -> bool

Returns true if the image pixels are a single channel (A8).

pub fn has_entries(&self) -> bool

If entries is not empty.

pub fn entries(&self) -> Vec<ImageVar>

Other images from the same container that are a child of this image.

pub fn flat_entries(&self) -> Var<Vec<(VarEq<ImageEntry>, usize)>>

All other images from the same container that are a descendant of this image.

The values are a tuple of each entry and the length of descendants entries that follow it.

The returned variable will update every time any entry descendant var updates.

pub fn entry_kind(&self) -> ImageEntryKind

Kind of image entry this image is in the source container.

pub fn entry_index(&self) -> usize

Sort index of the image in the list of entries of the source container.

pub fn with_best_reduce<R>( &self, size: PxSize, visit: impl FnOnce(&ImageEntry) -> R, ) -> Option<R>

Calls visit with the image or ImageEntryKind::Reduced entry that is nearest to size and greater or equal to it.

Does not call visit if none of the images are loaded, returns None in that case.

pub fn view_handle(&self) -> &ViewImageHandle

Connection to the image resource in the view-process.

pub fn layout_size(&self, ctx: &LayoutMetrics) -> PxSize

Calculate an ideal layout size for the image.

The image is scaled considering the density and screen scale factor. If the image has no density falls back to the screen_density in both dimensions.

pub fn calc_size( &self, ctx: &LayoutMetrics, fallback_density: PxDensity2d, ignore_image_density: bool, ) -> PxSize

Calculate a layout size for the image.

Parameters

ctx: Used to get the screen resolution.

fallback_density: Resolution used if density is None.

ignore_image_density: If true always uses the fallback_density as the resolution.

pub fn pixels(&self) -> Option<IpcBytes>

Reference the decoded pre-multiplied BGRA8 pixel buffer or A8 if is_mask.

pub fn partial_pixels(&self) -> Option<IpcBytes>

Reference the partially decoded pixels if the image is progressively decoding and has not finished decoding.

Format is BGRA8 for normal images or A8 if is_mask.

pub fn copy_pixels(&self, rect: PxRect) -> Option<(PxRect, IpcBytesMut)>

Copy the rect selection from pixels or partial_pixels.

The rect is in pixels, with the origin (0, 0) at the top-left of the image.

Returns the copied selection and the pixel buffer.

Note that the selection can change if rect is not fully contained by the image area.

pub async fn encode(&self, format: Txt) -> Result<IpcBytes, EncodeError>

Encode the image to the format.

Note that entries are ignored, only this image is encoded. Use encode_with_entries to encode multiple images in the same container.

pub async fn encode_with_entries( &self, entries: &[(ImageEntry, ImageEntryKind)], format: Txt, ) -> Result<IpcBytes, EncodeError>

Encode the images to the format.

This image is the first page followed by the entries in the given order.

pub async fn save(&self, path: impl Into<PathBuf>) -> Result<()>

Encode and write the image to path.

The image format is guessed from the file extension. Use save_with_format to specify the format.

Note that entries are ignored, only this image is encoded. Use save_with_entries to encode multiple images in the same container.

pub async fn save_with_format( &self, format: impl Into<Txt>, path: impl Into<PathBuf>, ) -> Result<()>

Encode and write the image to path.

The image is encoded to the format, the file extension can be anything.

Note that entries are ignored, only this image is encoded. Use save_with_entries to encode multiple images in the same container.

pub async fn save_with_entries( &self, entries: &[(ImageEntry, ImageEntryKind)], format: impl Into<Txt>, path: impl Into<PathBuf>, ) -> Result<()>

Encode and write the image to path.

The image is encoded to the format, the file extension can be anything.

This image is the first page followed by the entries in the given order.

pub fn insert_entry(&mut self, entry: ImageVar)

Insert entry in entries.

Parent Metadata

If the entry view-process metadata has another image as parent the metadata is replaced and a warning is logged.