Crate zng_ext_image

Image loading and cache.

Crate

This crate is part of the zng project.

Cargo Features

This crate provides 1 feature flag, not enabled by default.

"http"

Enable web image source.

Structs

IMAGES

Image loading, cache and render service.

IMAGES_WINDOW

Images render window hook.

IMAGE_RENDER

Controls properties of the render window used by IMAGES.render.

ImageHash

A 256-bit hash for image entries.

ImageHasher

Hasher that computes a ImageHash.

ImageLimits

Limits for image loading and decoding.

ImageManager

Application extension that provides an image cache.

ImagePpi

Pixels-per-inch of each dimension of an image.

ImageRenderArgs

Arguments for the ImageSource::Render closure.

Img

State of an ImageVar.

Enums

ImageCacheMode

Cache mode of IMAGES.

ImageDataFormat

Format of the image bytes.

ImageDownscale

Defines how an image is downscaled after decoding.

ImageMaskMode

Defines how the A8 image mask pixels are to be derived from a source mask image.

ImageSource

The different sources of an image resource.

ImageSourceFilter

Represents a PathFilter and UriFilter.

ProxyGetResult

Result of an ImageCacheProxy get redirect.

ProxyRemoveResult

Result of an ImageCacheProxy remove redirect.

Traits

ImageCacheProxy

A custom proxy in IMAGES.

ImageRenderWindowRoot

Implemented for the root window type.

ImageRenderWindowsService

Reference to a windows manager service that IMAGES can use to render images.

render_retain

Functions

render_retain

P If the render task is kept alive after a frame is produced, this is false by default meaning the image only renders once, if set to true the image will automatically update when the render widget requests a new frame.

Type Aliases

ImageVar

Represents an Img tracked by the IMAGES cache.

PathFilter

Represents a ImageSource::Read path request filter.

UriFilter

Represents a ImageSource::Download path request filter.