zng_ext_image

Trait render_retain

pub trait render_retain: WidgetExt {
    type MetaType;

    // Provided methods
    fn render_retain(&mut self, retain: impl IntoVar<bool>) { ... }
    fn unset_render_retain(&mut self) { ... }
}

Required Associated Types

type MetaType

Provided Methods

fn render_retain(&mut self, retain: impl IntoVar<bool>)

P If the render task is kept alive after a frame is produced, this is false by default meaning the image only renders once, if set to true the image will automatically update when the render widget requests a new frame.

This property sets and binds retain to IMAGE_RENDER.retain.

fn unset_render_retain(&mut self)

Unset the property.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.