zng_ext_image
Type Alias PathFilter

pub type PathFilter = ImageSourceFilter<PathBuf>;

Represents a ImageSource::Read path request filter.

Only absolute, normalized paths are shared with the Custom filter, there is no relative paths or .. components.

The paths are not canonicalized and existence is not verified, no system requests are made with unfiltered paths.

See ImageLimits::allow_path for more information.

Aliased Type

enum PathFilter {
    BlockAll,
    AllowAll,
    Custom(Arc<dyn Fn(&PathBuf) -> bool + Send + Sync>),
}

Variants

BlockAll

Block all requests of this type.

AllowAll

Allow all requests of this type.

Custom(Arc<dyn Fn(&PathBuf) -> bool + Send + Sync>)

Custom filter, returns true to allow a request, false to block.

Implementations

impl PathFilter

pub fn allow_dir(dir: impl AsRef<Path>) -> Self

Allow any file inside dir or sub-directories of dir.

pub fn allow_ext(ext: impl Into<OsString>) -> Self

Allow any path with the ext extension.

pub fn allow_current_dir() -> Self

Allow any file inside the env::current_dir or sub-directories.

Note that the current directory can be changed and the filter always uses the fresh current directory, use allow_dir to create a filter the always points to the current directory at the filter creation time.

pub fn allow_exe_dir() -> Self

Allow any file inside the current executable directory or sub-directories.

pub fn allow_res() -> Self

Allow any file inside the zng::env::res directory or sub-directories.